WIP: Enable Gateway sharing with TCP routing
See https://docs.google.com/document/d/1KGqdpXUPxTxvyOUih4ggmKBI5d8ePMF3_YLIvh28LfI

Add a PortRange field to GatewaySpec.Listeners - this is akin to a
listener template that causes listeners to be generated for every port
in the specified range.

Make Listener.Port optional, since it is not compatible with
Listener.PortRange.

Add the port number to ListenerStatus, to identify which port is
associated with each generated listener.